+++ lib/src/schema/field.dart
@@ -1,6 +1,7 @@
 import 'package:json_annotation/json_annotation.dart';
 import 'package:code_builder/code_builder.dart' as code_builder;
 import 'package:pocketbase_utils/src/utils/string_utils.dart';
+import 'package:pocketbase_utils/src/utils/utils.dart';
 
 part 'field.g.dart';
 
@@ -47,7 +48,7 @@ final class Field {
   code_builder.Reference fieldTypeRef(String className, {forceNullable = false}) {
     var fieldTypeRef = switch (type) {
       FieldType.text || FieldType.editor || FieldType.email || FieldType.url => 'String',
-      FieldType.number => 'int',
+      FieldType.number => options?.noDecimal == true ? 'int' : 'double',
       FieldType.bool => 'bool',
       FieldType.date => 'DateTime',
       FieldType.select => options?.maxSelect == 1 ? enumTypeName(className) : 'List<${enumTypeName(className)}>',
@@ -69,13 +70,41 @@ final class Field {
       ..modifier = code_builder.FieldModifier.final$
       ..type = fieldTypeRef(className));
   }
+
+  List<code_builder.Field> additionalFieldOptionsAsFields() {
+    return [
+      if (options?.min != null)
+        code_builder.Field((f) => f
+          ..static = true
+          ..modifier = code_builder.FieldModifier.constant
+          ..name = '${name}MinValue'
+          ..assignment = code_builder.Code(options!.min.toString())),
+      if (options?.max != null)
+        code_builder.Field((f) => f
+          ..static = true
+          ..modifier = code_builder.FieldModifier.constant
+          ..name = '${name}MaxValue'
+          ..assignment = code_builder.Code(options!.max.toString()))
+    ];
+  }
 }
 
 @JsonSerializable()
 class FieldOptions {
-  const FieldOptions({required this.maxSelect, required this.values});
+  const FieldOptions({
+    required this.min,
+    required this.max,
+    required this.noDecimal,
+    required this.maxSelect,
+    required this.values,
+  });
 
   final int? maxSelect;
+  @JsonKey(fromJson: jsonValueParseToInt)
+  final int? min;
+  @JsonKey(fromJson: jsonValueParseToInt)
+  final int? max;
+  final bool? noDecimal;
   final List<String>? values;
 
   factory FieldOptions.fromJson(Map<String, dynamic> json) => _$FieldOptionsFromJson(json);